chore: sc fmt

+++ docs/smart-contract-development/with-hardhat/contracts/Counter.sol
@@ -2,15 +2,17 @@
 pragma solidity ^0.8.4;
 
 contract Counter {
-    int private count = 0;
+    int256 private count = 0;
+
     function incrementCounter() public {
         count += 1;
     }
+
     function decrementCounter() public {
         count -= 1;
     }
 
-    function getCount() public view returns (int) {
+    function getCount() public view returns (int256) {
         return count;
     }
 }
